LayoutTests/fast/web-share/canShare-files.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<head>
    <title>PointerLock interface tests.</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../../resources/ui-helper.js"></script>
</head>
<body>
<div id="log"></div>
<script>
const textFile = new File(['hello'], 'hello.txt', {type:'text/plain'});
const emptyFile = new File([''], 'empty');

test(() => {
assert_throws_js(TypeError, () => { navigator.canShare({files: textFile}) });
}, 'canShare with single file');

test(() => {
assert_equals(navigator.canShare({files: []}), false);
}, 'canShare with empty file list');

test(() => {
assert_equals(navigator.canShare({files: [emptyFile]}), true);
}, 'canShare with single file list');

test(() => {
assert_equals(navigator.canShare({files: [textFile, emptyFile]}), true);
}, 'canShare with file list');

test(() => {
assert_equals(navigator.canShare({files: [textFile, emptyFile, textFile]}), true);
}, 'canShare with repeated file');

test(() => {
assert_equals(navigator.canShare({url: 'https://example.com/', files: [textFile, emptyFile]}), true);
}, 'canShare with file list and url');
</script>
</body>
</html>
