LayoutTests/fast/web-share/canShare-files-expected.txt


FAIL canShare with single file Can't find variable: assert_throws_js
PASS canShare with empty file list 
FAIL canShare with single file list assert_equals: expected true but got false
FAIL canShare with file list assert_equals: expected true but got false
FAIL canShare with repeated file assert_equals: expected true but got false
PASS canShare with file list and url 
